`KPlayerExoDrmCallback` now implements `ExtendedMediaDrmCallback` and provides an offline DrmSessionManager for downloaded media.
